fixed long long prob

+++ holberton.h
@@ -59,8 +59,8 @@ char *_utoa(unsigned int);
 /*octal.c */
 int conversion_o(char *s);
 char *make_octal(char *s, va_list vl);
-long long octorbi(unsigned int, int);
-char *_ltoa(long long);
+long octorbi(unsigned int, int);
+char *_ltoa(long);
 
 /*binary.c*/
 int conversion_b(char *s);
